+++ confik/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Optional `Source`s.
+- `FileSource` can be set as optional.
+
 ## 0.11.7
 
 - Implement `Configuration` for [`bigdecimal::BigDecimal`](https://docs.rs/bigdecimal/0.4/bigdecimal/struct.BigDecimal.html).