+++ confik/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## Unreleased
 
-- Minimum supported Rust version (MSRV) is now 1.67 due to `toml_edit` dependency.
+## 0.11.8
+
 - Implement `Configuration` for [`chrono::NaiveDateTime`](https://docs.rs/chrono/0.4/chrono/naive/struct.NaiveDateTime.html)
+- Minimum supported Rust version (MSRV) is now 1.67 due to transitive `toml_edit` dependency.
 
 ## 0.11.7